requirements.txt
# --- Data & Processing ---
pandas==2.1.4
numpy==1.26.2
xgboost>=2.0.0

# --- API Clients ---
facebook-business==18.0.0
spotipy==2.23.0
google-api-python-client==2.108.0

# --- Dashboard (Local only - Docker les installera mais ne s'en servira pas) ---
streamlit==1.29.0
plotly==5.18.0
watchdog==3.0.0
streamlit-authenticator==0.3.3
weasyprint>=60.0

# --- Utilities ---
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
cryptography>=42.0.0

# --- REST API (Brick 14) ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- Stripe billing (Brick 21) ---
stripe>=8.0.0

# --- Database Driver ---
psycopg2-binary==2.9.9

# --- Providers Airflow ---
# On ne force PAS la version de Airflow ici, car elle est déjà dans l'image Docker
# On ajoute juste les providers supplémentaires
apache-airflow-providers-postgres
apache-airflow-providers-http
apache-airflow-providers-common-sql
isodate==0.6.1

#Reading xlsx for meta_insight watcher
openpyxl
pandas
psycopg2-binary
python-dotenv